requirements.txt
warmup_scheduler==0.3.2
addict==2.2.1
pytorch_lightning==1.1.0rc1
openslide_python==1.1.2
albumentations==0.4.1
torch_optimizer==0.0.1a13
torchvision==0.8.2+cu110
pandas==0.25.1
opencv_python==4.2.0.34
segmentation_models_pytorch==0.1.3
matplotlib==3.1.1
tifffile==2021.1.11
pytorch_toolbelt==0.4.0
scikit_image==0.15.0
tqdm==4.36.1
torch==1.7.1+cu110
numpy==1.17.2
apex.egg==info
scipy==1.3.1
fastai.egg==info
apex==0.9.10dev
fastai==2.2.5
Pillow==8.1.0
PyYAML==5.3.1
scikit_learn==0.24.0
skimage==0.0
